Uncomment 'self.evict' calls in LRUCache
As evict can be an expensive call in the main thread, it was commented
out to check the improvement in performance. This commit uncomments it
for now. The code will be modified in later commits when async
eviction in a separate thread will be implemented.

Also uncomment the tests that were commented out earlier.