Refactor LRUCache code to support background eviction
This introduces new structs and enums to support background and
forground eviction in LRUCache. For now only foreground eviction is
supported which is exactly how it has been working so far. This commit
is backward compatible.

The field 'evict_strategy' of 'LRUCache' struct has been replaced with
'eviction_kind' which is the struct EvictionKind that encapsulates 3
things:

  1. eviction mode i.e. foreground v/s background

  2. strategy i.e deterministic v/s probabilistic

  3. trigger i.e. always v/s probabilistic

The 'Immediate' eviction strategy is renamed to 'Deterministic' as the
latter is a better description of the strategy, whereas the
'immediateness' of the strategy is indicated by the
'EvictionTrigger::Always' enum.

The field 'prob' in EvictionStrategy is now removed and instead the
one in EvictionTrigger::Probabilistic is considered.

The methods that execute the eviction are renamed to indicate exactly
what they do.