Add field pseudo_root in HNSWIndex struct
It's optional because pseudo nodes (including the pseudo root) will be
created in the index only if metadata schema is supported by the
index. This commit also includes changes to load pseudo root at the
time of loading the index.

The code for actually setting this field will be implemented in
subsequent commits.